requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.22.0
pandas>=1.5.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
Pillow>=9.2.0
tqdm>=4.64.0